Code refactoring: Added two separate services:
CareerProfileService
Including all Endpoints previously available in the UserService.
Separate testing for the service still needs to be implemented.

LearningHistoryService
Including all Endpoints previously available in the UserService
Separate testing for the service still needs to be implemented.

Added „Delete“ Endpoint for LearningHistory in the new LearningHistoryService.

Comment: The Endpoints for the CareerProfile and LearningHistory
are (currently) still also in the UserService. I will remove/migrate
the access to the Endpoints in the new two Services “CareerProfileService”
and “LearningHistoryService” as soon as possible.