rf(PathFinder): Code Cleanup: Removal of useless .catch(..)
Catch handles only NotFoundViolation, which is already handles in previous part -> Useless code
The only way to reach this code is in a hypothetical concurrent situation where an update of a path is handled while simultaneous the unit is deleted. A transaction should be used instead of a second superfluous check.